LangGraph/pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "httpx[socks]>=0.28.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langsmith>=0.4.10",
    "python-dotenv>=1.1.1",
]
